v0.3 - Test a new regex - not much in it
I was hoping that a different regex, one where it was matching non-input
characters would be faster - it would only need to find a non-input char
anywhere in the word. The old regex would need to match all the way to
the end for positive matches.

The benchmark script showed a very slight improvement 2%-ish for the
negative regex for long inputs, but no improvement for short inputs. On
reflection, the original regex would only have to read until the first
non-match as well, so the overall gain was tiny.

I used the new regex anyway.

+ Add benchmark script to test regex.
+ Update PSGI file to use new regex.